solve link song update not updated and update handling error when unlinked

+++ html/mp3online/protected/controllers/SongController.php
@@ -220,9 +220,13 @@ class SongController extends Controller
                 {
                     if(substr_count($oldLink,'http') == 0)
                     {
+                        $model->link = $linksong;                        
                         if(file_exists(Yii::app()->basePath.'/../upload/'.$oldLink))
                         {
-                            unlink(Yii::app()->basePath.'/../upload/'.$oldLink);
+                            if($oldLink)
+                            {
+                                unlink(Yii::app()->basePath.'/../upload/'.$oldLink);
+                            }
                         }                        
                     }
